requirements.txt
pyyaml
slackclient
websocket
websocket-client